Fixed serial input issue w/ asynPgpClSerial and ClSerialSlave classes.
With the Imperx camera, we were just getting the first 1 or 2 bytes of
the response and leaving the rest in the rogue input buffer.
This fix adds some diags and reworks the loop in ClSerialSlave to
pull bytes from the rogue input buffer as available and only timeout
once for the remaining bytes.